requirements.txt
fastapi
uvicorn
tree-sitter
lizard
pygithub
python-dotenv
requests
pyyaml
openai
python-multipart
ollama